[개발] 디테일 api 추가.

+++ lib/data/supabase/response/mission_clear_condition_response.dart
@@ -33,7 +33,7 @@ class MissionClearConditionResponse extends MissionClearConditionEntity {
   }) : super(
           id: id_.toInt(),
           mission: mission_,
-          count: count_,
+          count: count_.toInt(),
           title: title_,
           subtitle: subtitle_,
           content: content_,
